chore: unify comments style
