+++ internal/raft/errors.go
@@ -1,8 +1,10 @@
 package raft
 
+// Package raft implements a Raft consensus node and related components
+
 import "errors"
 
 var (
-	// ErrUnknownCommand is returned when an unknown command type is received
+	// ErrUnknownCommand is returned when trying to apply an unrecognized command type
 	ErrUnknownCommand = errors.New("unknown command type")
 )
