+++ internal/raft/fsm.go
@@ -9,25 +9,27 @@ import (
 	"github.com/vmihailenco/msgpack/v5"
 )
 
-// Command represents a command to be applied to the FSM
+// Command represents an operation to be replicated through the Raft cluster
 type Command struct {
-	Type    string        `json:"type"`
-	Key     string        `json:"key"`
-	Value   []byte        `json:"value"`
-	Timeout time.Duration `json:"timeout"`
+	Type    string        `json:"type"`    // Operation type (SET, DELETE, etc.)
+	Key     string        `json:"key"`     // Key to operate on
+	Value   []byte        `json:"value"`   // Value for SET operations
+	Timeout time.Duration `json:"timeout"` // Operation timeout
 }
 
-// fsm implements the raft.FSM interface
+// fsm implements the raft.FSM interface for state machine replication
 type fsm struct {
 	data map[string][]byte
 }
 
+// newFSM creates a new finite state machine instance
 func newFSM() *fsm {
 	return &fsm{
 		data: make(map[string][]byte),
 	}
 }
 
+// Apply handles incoming log entries and applies them to the state machine
 func (f *fsm) Apply(log *raft.Log) interface{} {
 	var cmd Command
 	if err := msgpack.Unmarshal(log.Data, &cmd); err != nil {
@@ -46,15 +48,18 @@ func (f *fsm) Apply(log *raft.Log) interface{} {
 	}
 }
 
+// Snapshot returns a snapshot of the current state machine
 func (f *fsm) Snapshot() (raft.FSMSnapshot, error) {
 	return &fsmSnapshot{data: f.data}, nil
 }
 
+// Restore restores the state machine from a snapshot
 func (f *fsm) Restore(rc io.ReadCloser) error {
 	f.data = make(map[string][]byte)
 	return json.NewDecoder(rc).Decode(&f.data)
 }
 
+// fsmSnapshot implements the FSMSnapshot interface for state persistence
 type fsmSnapshot struct {
 	data map[string][]byte
 }
